Refactor execution.
